requirements.txt
anyio==4.10.0
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ibm-cos-sdk==2.14.3
ibm-cos-sdk-core==2.14.3
ibm-cos-sdk-s3transfer==2.14.3
ibm_watsonx_ai==1.3.32
idna==3.10
jmespath==1.0.1
lomond==0.3.3
numpy==2.3.2
packaging==25.0
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
